[PS-38] Introduce a workflow that updates Trello when the main branch is updated (#10)
Changelog
====
- Introduces a workflow that updates Trello when the main branch is
updated. With this workflow, users don't need to update Trello manually
anymore.
- Introduces `change_card_to_done.sh`, a bash script that moves a card
that has a given ticket number in the name to the Done list.

Detailed Description
====
- In GitHub, we can add a workflow to our repository by creating a yaml
file in `.github/workflows`. We can set a condition that triggers the
workflow and GitHub automatically runs the workflow when the condition
is met.
- In this commit, the workflow is triggered when the main branch is
updated, and it runs `bin/change_card_to_done.sh`, passing a ticket
number in the main branch's latest commit message as an argument.
- Please note that you should add a ticket number in the title of the PR
if you want to change the status of the ticket as `Done` automatically.