[PS-59] Implement auto member assignment feature for bearcard (#24)
Changelog
====
- Implements auto member assignment feature for bearcard. Previously you
had to assign the member for each card in Trello manually but you don't
need to do once you run `setup-id-for-bearcard` after this PR is merged.
- Introduces `setup-id-for-bearcard`, a custom command that fetches the
Trello ID of a given name. If you run this command with your name used
in Trello, the ID will be fetched and it will be saved in
`~/.bearcard/id`.
- Fixes bearcard to fetch the ID from the file and send it to the API
end point when it create a card in Trello.
- Refactors bearcard to use `config.json`, instead of using a hard coded
API key & token.


Detailed Description
====
Please note that running `setup-id-for-bearcard` is optional. The card
will be managed by the administrator even if you don't run it.