[PS-64] Introduce a command that updates the Trello API key & token (#29)
Changelog
====
- Removes `command/config.json`, since it has raw API key & token and
leaving it in a remote repository can cause security risks.
- Introduces a command that updates the Trello API key & token by
generating a config file for it.
- Since the previous config file is removed in the repository with this
commit, it is necessary to a new command that generates the alternative
config file. You can generate it with this command.
- Changes the path of the config file for bearcard into
`~/.bearcard/config.json`. It is because the previous path was in the
repository but we decided to remove it.

Detailed Description
====
Please note that you should run `update-trello-api-key` if you want to
use `bearcard` after this commit is merged.